Add seccomp implementation
This finally adds support for all binaries. The previous ld_preload
implementation only worked if the binary is dynamically linked and links
to libc. If the binary was statically linked or called the syscall
directly without libc, then we would not detect the syscall.

Initially a plan was formed to add a ptrace implementation. While that
would work for all binaries, the disadvantage is that ptrace will have a
major performance impact (much like strace), since ptrace will be called
twice per syscall. Even then, it would be difficult to intercept the
syscall, as it is architecture dependent what register holds which
argument of the syscall. And most importantly it is very difficult to
overwrite pointer arguments, if the argument requires more space. For
example if one argument is a path of type char *, then if we would want
to replace the path with a longer string, we can't just assume that the
process has enough memory right at the original memory location. We
would painfully need to find a memory location that is safe to write.
This would be possible, but again architecture dependent and very
unelegant.
For example such an approach is used here [1].
In that case enough free memory is found on the stack, but this sounds
really unreliable to me and some care has to be taken to not collide
with a 128 bytes large red region on x86.

A much better approach (which this patch is implementing) is to use new
seccomp API [2].
The seccomp Notify API allows us to be notified for specific syscalls.
We can then respond to that syscall by either continuing the syscall
normally or by doing something completely different (including calling
the syscall with different arguments) and returning the result.
Since we are then calling the syscall from our own process, we are
completely free to allocate memory for syscall arguments in the normal
way.
Special care has to be taken when dealing with syscalls that return file
descriptors. Since we don't share address space with the target process,
the file descriptors will not be valid in the target process when we
return the result.
Luckily seccomp recently added brand new API to inject file descriptors
into other processes via SECCOMP_IOCTL_NOTIF_ADDFD [3].
By using seccomp we offload all the dirty architecture dependent work to
seccomp.

[1] https://www.alfonsobeato.net/c/modifying-system-call-arguments-with-ptrace/
[2] https://people.kernel.org/brauner/the-seccomp-notifier-new-frontiers-in-unprivileged-container-development
[3] https://people.kernel.org/brauner/the-seccomp-notifier-cranking-up-the-crazy-with-bpf