Add test for openat
Our original test for open() should actually not work, as we have not
implemented interception for open(2) yet.

It only works because the glibc interface for open() pretends to be
open(), but actually calls openat() behind the scenes.

Maybe we should in the future call the syscalls directly, the glibc
interface adds too much stuff on top of it.