Pass along the child exit status
For some strange reason we cannot use WNOHANG in the options for
waitpid, otherwise it will always return 0.
This might be another kernel bug with seccomp we would have to
investigate, as this does not make any sense since the child has
obviously exited.

However, with this we finally imitate the exit code of the supervised
process.