Add support for relative openat
For relative openat we need to duplicate the relative dirfd from the
supervised process (that is instead of injecting a fd, we need to
retrieve a fd in the opposite direction).

To accomplish this, we use pidfd_open() to get a pidfd for the child
early on and use the pidfd_getfd() function to duplicate a fd into the
supervisor.