Simplify BPF for loop
Instead of hardcoding the initial length and the ending length
constants, calculate the indices with an ever increasing i variable and
YOLO the length of the BPF program (but fingers crossed, MAX_FILTER_SIZE
should be large enough for the foreseeable future, right now we only
install 3 different syscall checks).

This should be much easier to understand than the previous index
shuffling.